separated all-view into its own reducer
