+++ src/store/configureStore.ts
@@ -9,11 +9,16 @@ import {
     RoadClosureAction,
     roadClosureReducer
 } from './road-closure';
+import {
+    RoadClosureExplorerAction,
+    roadClosureExplorerReducer
+} from './road-closure-explorer';
 
-export type RootAction = RoadClosureAction | ContextAction;
+export type RootAction = RoadClosureAction | RoadClosureExplorerAction | ContextAction;
 const rootReducer = combineReducers({
     context: contextReducer,
     roadClosure: roadClosureReducer,
+    roadClosureExplorer: roadClosureExplorerReducer,
 });
 export type RootState = StateType<typeof rootReducer>;
